automated_version/requirements.txt
fsspec
psycopg2-binary
pandas
matplotlib
shapely
logging
PyQt5